Moved PitchLookupModel to control communicator, deleted projectile_angle_convel, and removed unnecessary ROS2 parameters
