+++ src/prm_control/control_communicator/CMakeLists.txt
@@ -33,7 +33,7 @@ include_directories(include)
 add_executable(ControlCommunicatorNode 
   src/ControlCommunicatorNode.cpp
   src/ControlCommunicator.cpp
-  src/projectile_angle_convel.cpp
+  src/PitchLookupModel.cpp
   src/utils.cpp
 )
