+++ src/prm_vision/pnp_solver/include/PNPSolverNode.hpp
@@ -16,8 +16,6 @@
 #include "tf2/LinearMath/Quaternion.h"
 #include "tf2/LinearMath/Matrix3x3.h"
 
-#include "projectile_angle_convel.hpp"
-
 #include "PNPSolver.hpp"
 #include "KalmanFilter.hpp"
 #include "ValidityFilter.hpp"
